added about page with current version number
versioning scheme is going to be x.y where x is for feature additions or large feature updates and y is for small feature updates and bug fixes, etc.
This commit will be version 0.1 and will only be version 1.y when both Formation Updates and Package Merging for the Editor are added. Those features are necessary for Rank Up to be "complete" (emphasis on quotation marks).